Refactor browser cache stuff some.
