+++ fanficfare/browsercache/chromagnon/SuperFastHash.py
